Avoid running ldap backup on subclouds
LDAP should not be configured in subclouds. The bug being addressed
here caused LDAP to be erroneously be configured in controller-0 of
subclouds while not in controller-1. That caused confusion and different
testing behaviours when tests were performed against c0 and c1.

The script being fixed in this commit is one example of such
misunderstanding and confusion.

See other commits with this same topic for better comprehension.

Test cases:
PASS: In a subcloud with this change run
      '/etc/upgrade.d/11-ldap-users-backup.sh 21.12 22.12 start'.
      Verify that 'No actions required for this system type' is logged
      to /var/log/platform.log.
      Verify that 'echo $?' returns '0'.

Note: I don't think a full upgrade test is necessary for this change.
It looks straightforward to me and the testing I have performed already
ensures 'system ugprade-start' command will no longer fail because of
this script. Let me know if you disagree.

Partial-Bug: 2025663

+++ controllerconfig/controllerconfig/upgrade-scripts/11-ldap-users-backup.sh
@@ -30,6 +30,14 @@ log "$NAME: Saving backup of openldap schema files from release $FROM_RELEASE to
 
 if [[ "${ACTION}" == "start" ]] && [[ "${FROM_RELEASE}" == "21.12" ]] && [[ "${TO_RELEASE}" == "22.12" ]]; then
 
+    DISTRIBUTED_CLOUD_ROLE=$(source /etc/platform/platform.conf; \
+    echo $distributed_cloud_role)
+
+    if [[ $DISTRIBUTED_CLOUD_ROLE == "subcloud" ]] ; then
+        log "$NAME: No actions required for subclouds"
+        exit 0
+    fi
+
     BACKUP_DIR="/opt/platform/config/$FROM_RELEASE/ldap"
 
     rm -rf $BACKUP_DIR \
